+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # CyberArk AWS Cloud Package Release Notes
 
 The AWS Cloud package includes CyberArk PAS products, delivered as AMIs and AWS CloudFormation templates to automate deployment.
+
+## [10.8.1] (Release date 2019/01/23)
+
+### Added
+- AWS : Added - PTA Automatic deployment
+- AWS : Fix for - access PIMServices.svc failure
+- Azure : Added - Vault automatic deployment  
+
+
+## [10.7] (Release date 2019/01/23)
+
+### Added
+- Change deployment order to deploy the PVWA before CPM 
+### Removed
+- AIO template not released due to technical issue , will be released again in 10.8 
+
+
 ## [10.5] (Release date 2018/09/17)
 
 ### Added